+++ yrs/src/types/text.rs
@@ -167,7 +167,7 @@ impl TryFrom<Value> for TextRef {
 
 pub trait Text: AsRef<Branch> + Sized {
     /// Returns a number of characters visible in a current text data structure.
-    fn len<T: ReadTxn>(&self, txn: &T) -> u32 {
+    fn len<T: ReadTxn>(&self, txn: &T) -> u64 {
         self.as_ref().content_len
     }
 
@@ -219,7 +219,7 @@ pub trait Text: AsRef<Branch> + Sized {
     /// assert_eq!(ytext.get_string(txn), "Hi ★! to you");
     /// ```
     ///
-    fn insert(&self, txn: &mut TransactionMut, index: u32, chunk: &str) {
+    fn insert(&self, txn: &mut TransactionMut, index: u64, chunk: &str) {
         if chunk.is_empty() {
             return;
         }
@@ -251,7 +251,7 @@ pub trait Text: AsRef<Branch> + Sized {
     fn insert_with_attributes(
         &self,
         txn: &mut TransactionMut,
-        index: u32,
+        index: u64,
         chunk: &str,
         mut attributes: Attrs,
     ) {
@@ -280,7 +280,7 @@ pub trait Text: AsRef<Branch> + Sized {
     /// the end of it.
     ///
     /// This method will panic if provided `index` is greater than the length of a current text.
-    fn insert_embed<V>(&self, txn: &mut TransactionMut, index: u32, content: V) -> V::Return
+    fn insert_embed<V>(&self, txn: &mut TransactionMut, index: u64, content: V) -> V::Return
     where
         V: Into<EmbedPrelim<V>> + Prelim,
     {
@@ -308,7 +308,7 @@ pub trait Text: AsRef<Branch> + Sized {
     fn insert_embed_with_attributes<V>(
         &self,
         txn: &mut TransactionMut,
-        index: u32,
+        index: u64,
         embed: V,
         mut attributes: Attrs,
     ) -> V::Return
@@ -346,7 +346,7 @@ pub trait Text: AsRef<Branch> + Sized {
     /// Removes up to a `len` characters from a current text structure, starting at given `index`.
     /// This method panics in case when not all expected characters were removed (due to
     /// insufficient number of characters to remove) or `index` is outside of the bounds of text.
-    fn remove_range(&self, txn: &mut TransactionMut, index: u32, len: u32) {
+    fn remove_range(&self, txn: &mut TransactionMut, index: u64, len: u64) {
         let this = BranchPtr::from(self.as_ref());
         if let Some(pos) = find_position(this, txn, index) {
             remove(txn, pos, len)
@@ -357,7 +357,7 @@ pub trait Text: AsRef<Branch> + Sized {
 
     /// Wraps an existing piece of text within a range described by `index`-`len` parameters with
     /// formatting blocks containing provided `attributes` metadata.
-    fn format(&self, txn: &mut TransactionMut, index: u32, len: u32, attributes: Attrs) {
+    fn format(&self, txn: &mut TransactionMut, index: u64, len: u64, attributes: Attrs) {
         let this = BranchPtr::from(self.as_ref());
         if let Some(pos) = find_position(this, txn, index) {
             insert_format(this, txn, pos, len, attributes)
@@ -649,7 +649,7 @@ pub(crate) fn update_current_attributes(attrs: &mut Attrs, key: &str, value: &An
     }
 }
 
-fn find_position(this: BranchPtr, txn: &mut TransactionMut, index: u32) -> Option<ItemPosition> {
+fn find_position(this: BranchPtr, txn: &mut TransactionMut, index: u64) -> Option<ItemPosition> {
     let mut pos = {
         ItemPosition {
             parent: this.into(),
@@ -727,7 +727,7 @@ fn find_position(this: BranchPtr, txn: &mut TransactionMut, index: u32) -> Optio
     Some(pos)
 }
 
-fn remove(txn: &mut TransactionMut, mut pos: ItemPosition, len: u32) {
+fn remove(txn: &mut TransactionMut, mut pos: ItemPosition, len: u64) {
     let encoding = txn.store().options.offset_kind;
     let mut remaining = len;
     let start = pos.right.clone();
@@ -804,7 +804,7 @@ fn insert_format(
     this: BranchPtr,
     txn: &mut TransactionMut,
     mut pos: ItemPosition,
-    mut len: u32,
+    mut len: u64,
     attrs: Attrs,
 ) {
     minimize_attr_changes(&mut pos, &attrs);
@@ -1154,8 +1154,8 @@ impl TextEvent {
             action: Option<Action>,
             insert: Option<Value>,
             insert_string: Option<String>,
-            retain: u32,
-            delete: u32,
+            retain: u64,
+            delete: u64,
             attrs: Attrs,
             current_attrs: Attrs,
             delta: Vec<Delta>,
@@ -2195,7 +2195,7 @@ mod test {
 
             let c1 = text1.chars().count();
             let c2 = text2.chars().count();
-            let count = c1 as u32 + c2 as u32;
+            let count = c1 as u64 + c2 as u64;
 
             let _observer = text
                 .observe(move |txn, edit| assert_eq!(edit.delta(txn)[0], Delta::Deleted(count)));
@@ -2238,7 +2238,7 @@ mod test {
         let d2 = Doc::new();
         exchange_updates(&[&d1, &d2]);
 
-        txt.remove_range(&mut d1.transact_mut(), 0, "😭".len() as u32);
+        txt.remove_range(&mut d1.transact_mut(), 0, "😭".len() as u64);
         assert_eq!(txt.get_string(&txt.transact()).as_str(), "😊");
 
         exchange_updates(&[&d1, &d2]);
@@ -2256,7 +2256,7 @@ mod test {
         let d2 = Doc::new();
         exchange_updates(&[&d1, &d2]);
 
-        txt.remove_range(&mut d1.transact_mut(), 0, "⏰".len() as u32);
+        txt.remove_range(&mut d1.transact_mut(), 0, "⏰".len() as u64);
         assert_eq!(txt.get_string(&txt.transact()).as_str(), "⏳");
 
         exchange_updates(&[&d1, &d2]);
@@ -2272,7 +2272,7 @@ mod test {
         txt.insert(&mut txn, 0, "😊😭");
         // uncomment the following line will pass the test
         // txt.format(&mut txn, 0, "😊".len() as u32, HashMap::new());
-        txt.remove_range(&mut txn, "😊".len() as u32, "😭".len() as u32);
+        txt.remove_range(&mut txn, "😊".len() as u64, "😭".len() as u64);
 
         assert_eq!(txt.get_string(&txn).as_str(), "😊");
     }
@@ -2286,7 +2286,7 @@ mod test {
         txt.insert(&mut txn, 0, "⏰⏳");
         // uncomment the following line will pass the test
         // txt.format(&mut txn, 0, "⏰".len() as u32, HashMap::new());
-        txt.remove_range(&mut txn, "⏰".len() as u32, "⏳".len() as u32);
+        txt.remove_range(&mut txn, "⏰".len() as u64, "⏳".len() as u64);
 
         assert_eq!(txt.get_string(&txn).as_str(), "⏰");
     }
@@ -2301,11 +2301,11 @@ mod test {
 
         txt.format(
             &mut txn,
-            "👯".len() as u32,
-            "🙇‍♀️🙇‍♀️".len() as u32,
+            "👯".len() as u64,
+            "🙇‍♀️🙇‍♀️".len() as u64,
             HashMap::new(),
         );
-        txt.remove_range(&mut txn, "👯🙇‍♀️🙇‍♀️".len() as u32, "⏰".len() as u32); // will delete ⏰ and 👩‍❤️‍💋‍👨
+        txt.remove_range(&mut txn, "👯🙇‍♀️🙇‍♀️".len() as u64, "⏰".len() as u64); // will delete ⏰ and 👩‍❤️‍💋‍👨
 
         assert_eq!(txt.get_string(&txn).as_str(), "👯🙇‍♀️🙇‍♀️👩‍❤️‍💋‍👨");
     }
@@ -2320,13 +2320,13 @@ mod test {
         txt.insert(&mut txn, 0, "👯");
         txt.format(
             &mut txn,
-            "👯".len() as u32,
-            "🙇‍♀️🙇‍♀️".len() as u32,
+            "👯".len() as u64,
+            "🙇‍♀️🙇‍♀️".len() as u64,
             HashMap::new(),
         );
 
         // will delete ⏰ and 👩‍❤️‍💋‍👨
-        txt.remove_range(&mut txn, "👯🙇‍♀️🙇‍♀️".len() as u32, "⏰".len() as u32); // will delete ⏰ and 👩‍❤️‍💋‍👨
+        txt.remove_range(&mut txn, "👯🙇‍♀️🙇‍♀️".len() as u64, "⏰".len() as u64); // will delete ⏰ and 👩‍❤️‍💋‍👨
 
         assert_eq!(&txt.get_string(&txn), "👯🙇‍♀️🙇‍♀️👩‍❤️‍💋‍👨");
     }
@@ -2341,21 +2341,21 @@ mod test {
         txt.insert(&mut txn, 0, "👯");
         txt.format(
             &mut txn,
-            "👯".len() as u32,
-            "❤️❤️🙇‍♀️🙇‍♀️⏰".len() as u32,
+            "👯".len() as u64,
+            "❤️❤️🙇‍♀️🙇‍♀️⏰".len() as u64,
             HashMap::new(),
         );
-        txt.insert(&mut txn, "👯❤️❤️🙇‍♀️🙇‍♀️⏰".len() as u32, "⏰");
+        txt.insert(&mut txn, "👯❤️❤️🙇‍♀️🙇‍♀️⏰".len() as u64, "⏰");
         txt.format(
             &mut txn,
-            "👯❤️❤️🙇‍♀️🙇‍♀️⏰⏰".len() as u32,
-            "👩‍❤️‍💋‍👨".len() as u32,
+            "👯❤️❤️🙇‍♀️🙇‍♀️⏰⏰".len() as u64,
+            "👩‍❤️‍💋‍👨".len() as u64,
             HashMap::new(),
         );
         txt.remove_range(
             &mut txn,
-            "👯❤️❤️🙇‍♀️🙇‍♀️⏰⏰👩‍❤️‍💋‍👩".len() as u32,
-            "👩‍❤️‍💋‍👨".len() as u32,
+            "👯❤️❤️🙇‍♀️🙇‍♀️⏰⏰👩‍❤️‍💋‍👩".len() as u64,
+            "👩‍❤️‍💋‍👨".len() as u64,
         );
         assert_eq!(txt.get_string(&txn).as_str(), "👯❤️❤️🙇‍♀️🙇‍♀️⏰⏰👩‍❤️‍💋‍👨");
     }
